+++ src/main/resources/db/migration/V1__init.sql
@@ -1,19 +1,19 @@
 CREATE TABLE users (
-    id SERIAL PRIMARY KEY,
+    id BIGSERIAL PRIMARY KEY,
     email VARCHAR(255) NOT NULL,
     name VARCHAR(255) NOT NULL,
     surname VARCHAR(255) NOT NULL
 );
 
 CREATE TABLE accounts (
-    id SERIAL PRIMARY KEY,
+    id BIGSERIAL PRIMARY KEY,
     name VARCHAR(255) NOT NULL,
     type VARCHAR(255),
     industry VARCHAR(255)
 );
 
 CREATE TABLE user_accounts (
-    user_id INTEGER NOT NULL REFERENCES users(id),
-    account_id INTEGER NOT NULL REFERENCES accounts(id),
+    user_id BIGINT NOT NULL REFERENCES users(id),
+    account_id BIGINT NOT NULL REFERENCES accounts(id),
     PRIMARY KEY (user_id, account_id)
 );
